setup the Homepage route
