+++ client/src/components/App.js
@@ -1,23 +1,42 @@
 import React from 'react';
 import './App.scss';
-import useApplicationData from '../hooks/useApplicationData'
 
-function App() {
-  const { state, dispatch } = useApplicationData();
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link
+} from "react-router-dom";
+import Homepage from './Homepage';
 
-  const userList = state.users.map(user => (
-    <li key={user.id}>
-      {user.first_name} {user.last_name} {user.email}
-    </li>
-  ));
-  
+export default function App() {
+
   return (
-    <div className="App">
-      <h1>Users</h1>
+    <Router>
+      <div>
+        <nav>
+          <ul>
+            <li>
+              <Link to="/">Home</Link>
+            </li>
+            <li>
+              <Link to="/cities/1">City</Link>
+            </li>
+            <li>
+              <Link to="/users/1">Users</Link>
+            </li>
+            <li>
+              <Link to="/trips/1">Trip</Link>
+            </li>
+          </ul>
+        </nav>
 
-      <ul>{userList}</ul>
-    </div>
+        <Switch>
+          <Route exact path="/">
+            <Homepage />
+          </Route>
+        </Switch>
+      </div>
+    </Router>
   );
-}
-
-export default App;
+};
